pawan/spider/spider/requirements.txt
redis==3.5.3
requests==2.25.1
PyYAML==5.4.1
celery==5.2.1
mysql-connector-python == 8.0.27
kafka-python==2.0.2
google-cloud-storage==1.38.0
boto3==1.17.67
botocore==1.20.106
